Yacc(Updated)
The Yaccc file was updated with the expressions and calculations now working. The word input doesn't fully work however. If you add an expression after the words with any recognized calculation type then it calculates it and gives the answer. But it still gives the syntax error, I think because of its placement in the code. If you change the placements, more errors might occur.